cmd/geth, metclient: fix 32-bit builds
Two spots did not compile under GOARCH=386, found when the i386 workflow
ran against the v1.13.14 tree for the first time:

- the default gas price (80000000000 wei) does not fit a 32-bit int.
  Widen the gasprice flag and the Deploy/SendValue price parameters to
  int64; wei amounts never belonged in a platform-sized int.
- syscall.Rusage.Maxrss is int32 on 32-bit platforms; compare through an
  explicit int64 conversion.

+++ cmd/geth/governancedeploy.go
@@ -88,7 +88,7 @@ func deployGovernanceContracts(cliCtx *cli.Context) error {
 	// get command line arguments
 	url := cliCtx.String(urlFlag.Name)
 	gas := cliCtx.Int(gasFlag.Name)
-	gasPrice := cliCtx.Int(gasPriceFlag.Name)
+	gasPrice := cliCtx.Int64(gasPriceFlag.Name)
 
 	if gas <= 0 {
 		gas = 0xF000000
